Handle mepo-cd failure gracefully
The issue was that if you did `mepo-cd blabla` and blabla isn't a
component, it would execute `cd` and then dump you into your home
directory.

Instead, if we `raise SystemExit` in the `verify` we can trap a non-zero
status code and stay in place in case of an error.